fix(zano): close reconciliation gaps found in fix verification
A verification pass on the previous commit confirmed the reopen removal
and tri-state probe but found the remaining evidence and signalling gaps
below. All are addressed here rather than deferred.

Two-source agreement before unaccounting: a daemon -14 alone no longer
counts as a miss. The wallet is asked directly (get_bulk_payments for the
payment id) and must also no longer list the transaction; if the two
endpoints diverge the payment is left untouched and the divergence is
logged. Daemon and wallet are independently configured nodes and either
one alone can be stale.

Honest recovery: a payment restored from Unaccounted has its chain
placement recomputed from the daemon's tx_info.keeper_block (-1 = pool
only -> zero confirmations, Processing at most; >0 = confirmed at that
height -> confirmations from the current tip). The pre-drop confirmation
count is never reused, and recovery never reads from the probe cache.

Merchant-facing signal: a store-scoped dashboard notification
(ZanoPaymentReconciliationNotification: PaymentLost /
ConfirmationsRegressed / PaymentRestored) is sent on each transition,
de-duplicated by (kind, invoice, tx). The critical log is now
supplemental rather than the only signal.

Settlement-anchored window: ZanoPaymentData.SettledAt records the first
time a payment reached Settled and the 48h reconciliation window is
measured from it (rows predating the field fall back to Created).
Candidates are pre-filtered in SQL by Created over a 14-day look-back and
the anchor is applied in memory from the payment blob.

Boundary handling: invoices with an in-flight sub-threshold miss sequence
stay selectable past the window end (in-memory sticky set, 1h TTL
backstop), and cached daemon positives are bypassed in the last
(cache TTL + threshold polls) of the window.

Late payments to settled invoices are recorded again (idempotent by id)
instead of dropped: real receipts must appear in payment history, API
and refund flows even when the invoice is terminal.

RPC: get_bulk_payments added to the fast-timeout and retry allow-lists;
its amount field widened to decimal; get_tx_details tx_info typed to
expose keeper_block.

Tests: ClassifyTxDetailsResponse matrix updated for the typed tx_info;
ApplyRecoveredPlacement, IsWithinReconciliationWindow and
IsInReconciliationTail covered. 119/119 pass.

+++ Plugins/Zano/RPC/Models/GetBulkPaymentsResponse.cs
@@ -11,7 +11,7 @@ namespace BTCPayServer.Plugins.Zano.RPC.Models
 
     public class Payment
     {
-        [JsonProperty("amount")] public long Amount { get; set; }
+        [JsonProperty("amount")] public decimal Amount { get; set; }
         [JsonProperty("block_height")] public long BlockHeight { get; set; }
         [JsonProperty("payment_id")] public string PaymentId { get; set; }
         [JsonProperty("tx_hash")] public string TxHash { get; set; }
